plugin-system: More separation of api and logic
